requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
Pillow==8.4.0
psycopg2==2.9.2
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
